feat: add main page
