+++ src/Pages/Main/MainPage.module.scss
@@ -1,0 +1,6 @@
+.main {
+  max-width: 70%;
+  margin: auto;
+  display: flex;
+  justify-content: center;
+}
