+++ src/Pages/Main/MainPage.tsx
@@ -1,10 +1,11 @@
+import { UserCards } from '../../components/UserCards/UserCards';
 import styles from './MainPage.module.scss';
 
 export function MainPage() {
   return (
     <>
       <main className={styles.main}>
-        <p>Main Page</p>
+        <UserCards />
       </main>
     </>
   );
