+++ src/components/UserCard/UserCard.module.scss
@@ -1,0 +1,12 @@
+.CardContainer {
+  display: flex;
+  flex-direction: column;
+  border: 1px solid black;
+  border-radius: 5px;
+  padding: 10px;
+}
+
+.CardBlock {
+  display: flex;
+  gap: 5px;
+}
