+++ src/components/UserCards/UserCards.module.scss
@@ -1,0 +1,7 @@
+.cardsContainer {
+  width: 100%;
+  display: flex;
+  flex-wrap: wrap;
+  gap: 10px;
+  justify-content: center;
+}
